move get_mjd() to utils.py
